Ground Enemy
Removed the AI debug messages
Added the Enemy script and the Photon View to the Ground Enemy
Modified the LevelLoader.cs to add a GroundEnemy to the load.
GroundEnemy now is working as an enemy, can be killed and stuff. Also
networking should be workinh with him (I didn't test).